建置GooglePlaceAPI 功能 GetNearbyPlace  完成。 建立Trip 模組 完成

+++ prjFriendlyFoodWebAPI/Services/FoodMap/Interfaces/IRecommendationServices.cs
@@ -5,6 +5,6 @@ namespace prjFriendlyFoodWebAPI.Services.FoodMap.Interfaces
 {
     public interface IRecommendationServices
     {
-        Task<List<RecommendationDTO>> GetRecommendationDTOsAsync(long shoppinglistId); 
+        Task<List<RecommendationDTO>> GetRecommendationDTOsAsync(int shoppinglistId); 
     }
 }
